microgames/pyproject.toml
[tool.poetry]
name = "microgames"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11, <4.0"
grpcio = "^1.53.0"
grpcio-tools = "^1.53.0"
requests = "^2.28.2"
redis = {extras = ["hiredis"], version = "^4.5.4"}
pymongo = "^4.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"